Updated the code to send across the csv file
