+++ admin/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "start": "node src/index.js",
     "develop": "nodemon src/index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha"
   },
   "repository": {
     "type": "git",
@@ -24,6 +24,7 @@
   "devDependencies": {
     "@mft/eslint-config-momentumft": "^10.0.0",
     "eslint": "^8.56.0",
+    "mocha": "^10.2.0",
     "nodemon": "^2.0.7"
   },
   "dependencies": {
@@ -31,6 +32,7 @@
     "body-parser": "^1.19.0",
     "config": "^3.3.3",
     "express": "^4.17.1",
+    "json2csv": "^6.0.0-alpha.2",
     "ramda": "^0.27.1",
     "request": "^2.88.2"
   }
